Conflicts:
	plugins/de.cau.cs.kieler.klay.planar/src/de/cau/cs/kieler/klay/planar/graph/package-info.java
	plugins/de.cau.cs.kieler.klay.planar/src/de/cau/cs/kieler/klay/planar/intermediate/package-info.java
	plugins/de.cau.cs.kieler.klay.planar/src/de/cau/cs/kieler/klay/planar/p1planar/package-info.java
	plugins/de.cau.cs.kieler.klay.planar/src/de/cau/cs/kieler/klay/planar/p2ortho/package-info.java
removed package info local and from index

+++ plugins/de.cau.cs.kieler.klay.layered/src/de/cau/cs/kieler/klay/layered/p5edges/SplineEdgeRouter.java
@@ -124,7 +124,7 @@ public class SplineEdgeRouter extends AbstractAlgorithm implements ILayoutPhase 
         layeredGraph.getSize().x = xpos - spacing;
 
         // get user defined minimal angle for straight edges heading in and out nodes.
-        int minimalAngle = layeredGraph.getProperty(Properties.MIN_EDGE_ANGLE);
+        float minimalAngle = layeredGraph.getProperty(Properties.MIN_EDGE_ANGLE);
         // check all short edges
         if (minimalAngle != 0) {
             for (LEdge edge : shortEdges) {
